move global dispatcher to extension package
